Feat: 건강 정보 기록 시 정수로 입력하도록 수정

+++ pillsogood-server/src/graphql/health/health.typeDefs.ts
@@ -3,10 +3,10 @@ import { gql } from 'apollo-server-express';
 export default gql`
     type Health {
         _id:String,
-        height:Float,
-        weight:Float,
-        hypertension:Float,
-        bloodSugarLevel:Float,
+        height:Int,
+        weight:Int,
+        hypertension:Int,
+        bloodSugarLevel:Int,
         createdAt:String
     }
 
@@ -15,7 +15,7 @@ export default gql`
     }
 
     type Mutation {
-        createHealthRecord(jwt:String!, height:Float!, weight:Float!, hypertension:Float!, bloodSugarLevel:Float!):Int!
+        createHealthRecord(jwt:String!, height:Int!, weight:Int!, hypertension:Int!, bloodSugarLevel:Int!):Int!
         updateHealthRecord(jwt:String!, _id:String!, weight:Int!, hypertension:Int!, bloodSugarLevel:Int!):Int!
         deleteHealthRecord(jwt:String!, _id:String!):Int!
     }
